+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 - Breaking changes:
   - Removed `Skin::has_cursor_middle` and `Render::has_cursor_middle` ([#2])
   - Added the fields `RenderOptions::show_strain_graph`, `RenderOptions::show_slider_breaks`, and `RenderOptions::ignore_fail`
+  - Removed `PartialEq` and `Eq` impls for `RenderList` and `Render`
+  - The field `Render::skin` is now of type `RenderSkinOption<'static>`
+  - Added the field `Render::options`
 
 - Additions:
   - Implement `Default` for `OrdrClient`